+++ test/libsolidity/syntaxTests/empty_struct.sol
@@ -1,0 +1,5 @@
+contract test {
+	struct A {}
+}
+// ----
+// Warning: Defining empty structs is deprecated.